backend/pyproject.toml
[tool.poetry]
name = "prophy"
version = "0.1.0"
description = "Aplicação Web para atendimentos de clientes no setor da física médica"
authors = ["ledsouza <leandro.souza.159@gmail.com>"]
license = "GPL-3.0-only"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.2"
djangorestframework = "^3.15.2"
Markdown = "^3.6"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.5.1"
djoser = "^2.3.0"
pillow = "^12.1.1"
django-stubs = "^5.0.4"
localflavor = "^1.9"
faker = "^28.1.0"
validate-docbr = "^1.10.0"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
setuptools = "^78.1.1"
aiosmtpd = "^1.4.6"
django-ses = "^4.3.0"
mailgun = "^1.1.0"
django-anymail = {extras = ["mailgun"], version = "^13.0.1"}
google-auth = "^2.40.3"
reportlab = "^4.4.3"
pydantic = "^2.11.7"
phonenumbers = "^9.0.12"
django-storages = {extras = ["google"], version = "^1.14.6"}
psycopg = {extras = ["binary"], version = "^3.2.10"}
whitenoise = "^6.10.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
pytest-xdist = "^3.8.0"
pytest-mock = "^3.15.1"
factory-boy = "^3.3.1"
django-cypress = "^1.0.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.test"
pythonpath = ["."]
testpaths = ["."]
addopts = "-ra --reuse-db --cov=. --cov-report=term-missing"

[tool.coverage.run]
source = ["."]

[tool.coverage.report]
fail_under = 0
show_missing = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
